+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 -   Introduce Azure CLI
 -   Install Azure CLI resource-graph extension if it is not installed
 -   Enable filtering of Azure infrastructure resources, e.g. virtual machines
+-   Enable filtering of Azure infrastructure resources in probes
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-azure/compare/0.2.0...HEAD